[EDIT] remove usb echi build warnings

+++ arch/arm/cpu/slsiap/devices/ehci-hcd-hsic.c
@@ -740,7 +740,7 @@ ehci_submit_root(struct usb_device *dev, unsigned long pipe, void *buffer,
 			if (HCS_PPC(ehci_readl(&ctrl->hccr->cr_hcsparams))) {
 				reg |= EHCI_PS_PP;
 				ehci_writel(status_reg, reg);
-				if( status_reg == 0xc0030058 )
+				if( (unsigned long)status_reg == 0xc0030058 )
 				{
 					reg2 |= 0x02;
 					ehci_writel(status_reg2, reg2);
